Optionally allow additional CLI arguments to start
This commit adds a new parameter, 'other_args' which
can be used to pass additional parameters to the start
command of a process.

This is particularly useful with clustered software such as
consul or mysql-galera, where the initial service creation
requires a bootstrapping, but once the cluster has been
started it can be treated the same as every other service.